refactor: change ValidateParameters and GetFileFromHeader to httputils and your refactor calls in all classes :hammer:
