+++ common/net/http/headers.go
@@ -5,4 +5,6 @@ const (
 	headerUserAgent     = "User-Agent"
 	headerRealIP        = "X-Real-Ip"
 	headerForwardedFor  = "X-Forwarded-For"
+	dsl                 = "dsl"
+	fileExtension       = ".gold"
 )
