+++ components/ledger/internal/domain/metadata/metadata_repository.go
@@ -3,7 +3,7 @@ package metadata
 import (
 	"context"
 
-	"github.com/LerianStudio/midaz/common"
+	commonHTTP "github.com/LerianStudio/midaz/common/net/http"
 )
 
 // Repository provides an interface for operations related on mongodb a metadata entities.
@@ -11,7 +11,7 @@ import (
 //go:generate mockgen --destination=../../gen/mock/metadata/metadata_mock.go --package=mock . Repository
 type Repository interface {
 	Create(ctx context.Context, collection string, metadata *Metadata) error
-	FindList(ctx context.Context, collection string, filter common.QueryHeader) ([]*Metadata, error)
+	FindList(ctx context.Context, collection string, filter commonHTTP.QueryHeader) ([]*Metadata, error)
 	FindByEntity(ctx context.Context, collection, id string) (*Metadata, error)
 	Update(ctx context.Context, collection, id string, metadata map[string]any) error
 	Delete(ctx context.Context, collection, id string) error
